admin_panel_service/sqlite_to_postgres/requirements.txt
psycopg2-binary==2.9.7
pytest==7.4.2
python-dotenv==1.0.0
flake8==6.1